+++ src/consensus/ledger_enclave.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "consensus/ledger_enclave_types.h"
 #include "ds/serialized.h"
 #include "kv/kv_types.h"
