+++ src/ds/champ_map.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "ds/buffer.h"
 #include "ds/hash.h"
 #include "ds/map_serializers.h"
