+++ src/kv/generic_serialise_wrapper.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "ds/buffer.h"
 #include "kv_types.h"
 #include "node/rpc/claims.h"
