+++ src/node/historical_queries.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "ccf/historical_queries_interface.h"
 #include "consensus/ledger_enclave_types.h"
 #include "kv/store.h"
