+++ src/node/request_tracker.h
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the Apache 2.0 License.
 #pragma once
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "crypto/hash.h"
 #include "ds/dl_list.h"
