+++ src/node/snapshotter.h
@@ -2,7 +2,7 @@
 // Licensed under the Apache 2.0 License.
 #pragma once
 
-#include "ccf/assert.h"
+#include "ccf/ccf_assert.h"
 #include "ccf/ds/logger.h"
 #include "consensus/ledger_enclave_types.h"
 #include "crypto/hash.h"
